Move the AWS SSO login module out of credentials/
It spawns a subprocess and classifies errors, which is re-authentication
machinery rather than credential resolution. Every other file under
credentials/ is free of vscode imports and user-facing strings; this keeps
it that way, and matches how the Databricks OAuth machinery sits at the
src root alongside credentials/databricks.ts.

See #15272
